Use proper size_t type for how much bytes are allocated

+++ vm/objectmemory.cpp
@@ -350,7 +350,7 @@ namespace rubinius {
     // std::cout << "Pruned " << count << " handles, " << total << "/" << handles->size() << " total.\n";
   }
 
-  int ObjectMemory::mature_bytes_allocated() {
+  size_t ObjectMemory::mature_bytes_allocated() {
     return immix_->bytes_allocated() + mark_sweep_->allocated_bytes;
   }
